+++ packages/sdk/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Unreleased
 
+- add more tests for new `NotesDB` impl
 - fix db tests
-- make `NotesManager` use `removeNotesByNullifier` instead of looping through the entire DB
+- make `NotesManager` use `removeNotesByNullifiers` instead of looping through the entire DB
 - re-implement `NotesDB` to index notes by `merkleIndex`, `nullifier`, and `asset`. 
 - rename `getNotesFor` -> `getNotesForAsset`
 - make `OpPreparer` and `NotesManager` use `NocturneViewer` instead of `NocturneSigner`